Extract utility methods
