+++ lib/resolve-eslint-config.js
@@ -5,6 +5,12 @@ const path = require('path')
 
 const pkgConf = require('pkg-conf')
 
+const {
+  ensureArray,
+  ensureStringArrayValue,
+  stringArrayToObj
+} = require('./utils')
+
 /** @typedef {import('../').EslintOptions} EslintOptions */
 
 const DEFAULT_EXTENSIONS = [
@@ -20,47 +26,6 @@ const DEFAULT_IGNORE = [
   'node_modules/**',
   'vendor/**'
 ]
-
-/**
- * @template T
- * @param {string|string[]} values
- * @param {{ [key: string]: T }} base
- * @returns {{ [key: string]: T|true }}
- */
-const stringArrayToObj = (values, base = {}) => {
-  /** @type {{ [key: string]: T|true }} */
-  const result = { ...base }
-
-  for (const value of ensureArray(values)) {
-    result[value] = true
-  }
-
-  return result
-}
-
-/**
- * @template T
- * @param {T[]|T} value
- * @returns {T[]}
- */
-const ensureArray = (value) => Array.isArray(value) ? [...value] : [value]
-
-/**
- * @param {unknown} value
- * @returns {string[]}
- */
-const ensureStringArrayValue = (value) => {
-  if (!Array.isArray(value)) return []
-
-  /** @type {string[]} */
-  const result = []
-
-  for (const item of value) {
-    if (typeof item === 'string') result.push(item)
-  }
-
-  return result
-}
 
 /**
  * @param {EslintOptions} eslintConfig
